Refactored kinderzuschlag.
