+++ gettsim/tests/test_data/test_dfs_tax_transfer.csv
@@ -1,15 +1,15 @@
-hh_id,tu_id,p_id,tu_vorstand,anz_erwachsene_hh,anz_minderj_hh,vermögen_hh,bruttolohn_m,wohnort_ost,alter,selbstständig,hat_kinder,eink_selbstst_m,ges_rente_m,prv_krankenv_beitr_m,bruttolohn_vorj_m,arbeitsl_lfdj_m,arbeitsl_vorj_m,arbeitsl_vor2j_m,arbeitsstunden_w,anz_kinder_tu,anz_erw_tu,geburtsjahr,geburtstag,geburtsmonat,m_elterngeld_mut,m_elterngeld_vat,m_elterngeld,exper,entgeltpunkte,kind,rentner,betreuungskost_m,prv_rente_beitr_m,miete_unterstellt,kapital_eink_m,vermiet_eink_m,jahr_renteneintr,gem_veranlagt,in_ausbildung,behinderungsgrad,alleinerziehend,kaltmiete_m,wohnfläche,bewohnt_eigentum,heizkost_m,immobilie_baujahr,mietstufe,sonstig_eink_m,jahr
-1,1,1,TRUE,4,0,5500,300,FALSE,25,FALSE,FALSE,0,0,FALSE,1000,1,0,0,0,0,4,1958,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,0,2061,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
-1,1,2,FALSE,4,0,5500,600,FALSE,25,FALSE,FALSE,0,0,FALSE,2000,1,0,0,0,0,4,1957,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,1,2062,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
-1,2,3,FALSE,4,0,5500,900,FALSE,25,FALSE,FALSE,0,0,FALSE,3000,12,5,0,0,1,4,1982,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,2,2063,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
-1,2,4,FALSE,4,0,5500,1200,TRUE,25,FALSE,TRUE,0,0,FALSE,4000,1,0,0,20,1,4,1970,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,3,2064,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
-2,3,5,TRUE,4,0,5500,1500,TRUE,25,FALSE,TRUE,0,0,FALSE,7000,1,0,0,0,1,4,1952,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,4,2065,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
-2,3,6,FALSE,4,0,5500,5000,TRUE,25,FALSE,TRUE,0,0,FALSE,2500,1,0,0,0,1,4,1985,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,201,5,2066,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
-2,4,7,FALSE,4,0,5500,300,TRUE,20,FALSE,FALSE,0,0,FALSE,2300,1,0,0,0,0,4,1982,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,202,6,2067,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
-2,4,8,FALSE,4,0,5500,600,TRUE,20,FALSE,FALSE,0,0,FALSE,1000,1,0,0,0,0,4,1971,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,203,7,2068,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
-3,5,9,TRUE,1,0,5500,900,TRUE,20,FALSE,FALSE,0,0,FALSE,2000,1,0,0,0,0,1,1968,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,204,8,2069,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
-4,6,10,TRUE,1,0,5500,1200,FALSE,20,FALSE,TRUE,0,0,FALSE,3000,12,5,0,0,1,1,1964,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,205,9,2070,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
-5,7,11,TRUE,1,0,5500,1500,FALSE,20,FALSE,TRUE,0,0,FALSE,4000,1,0,0,20,1,1,1967,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,206,10,2071,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
-6,8,12,TRUE,1,0,5500,8000,FALSE,20,FALSE,TRUE,0,0,FALSE,7000,1,0,0,0,1,1,1951,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,207,11,2072,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
-7,9,13,TRUE,1,0,5500,700,FALSE,20,FALSE,FALSE,0,0,FALSE,2500,1,0,0,0,1,1,1960,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,208,12,2073,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2002
-8,10,14,TRUE,1,0,5500,0,FALSE,40,TRUE,TRUE,2500,0,FALSE,2300,1,0,0,0,0,1,1959,1,1,0,0,0,22,19.7296,FALSE,FALSE,0,0,0,0,13,2074,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2019
+hh_id,tu_id,p_id,tu_vorstand,anz_minderj_hh,vermögen_hh,bruttolohn_m,wohnort_ost,alter,selbstständig,hat_kinder,eink_selbstst_m,ges_rente_m,prv_krankenv_beitr_m,bruttolohn_vorj_m,arbeitsl_lfdj_m,arbeitsl_vorj_m,arbeitsl_vor2j_m,arbeitsstunden_w,anz_kinder_tu,anz_erw_tu,geburtsjahr,geburtstag,geburtsmonat,m_elterngeld_mut,m_elterngeld_vat,m_elterngeld,exper,entgeltpunkte,kind,rentner,betreuungskost_m,prv_rente_beitr_m,miete_unterstellt,kapital_eink_m,vermiet_eink_m,jahr_renteneintr,gem_veranlagt,in_ausbildung,behinderungsgrad,alleinerziehend,kaltmiete_m,wohnfläche,bewohnt_eigentum,heizkost_m,immobilie_baujahr,mietstufe,sonstig_eink_m,jahr
+1,1,1,TRUE,0,5500,300,FALSE,25,FALSE,FALSE,0,0,FALSE,1000,1,0,0,0,0,4,1958,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,0,2061,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
+1,1,2,FALSE,0,5500,600,FALSE,25,FALSE,FALSE,0,0,FALSE,2000,1,0,0,0,0,4,1957,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,1,2062,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
+1,2,3,FALSE,0,5500,900,FALSE,25,FALSE,FALSE,0,0,FALSE,3000,12,5,0,0,1,4,1982,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,2,2063,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
+1,2,4,FALSE,0,5500,1200,TRUE,25,FALSE,TRUE,0,0,FALSE,4000,1,0,0,20,1,4,1970,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,3,2064,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
+2,3,5,TRUE,0,5500,1500,TRUE,25,FALSE,TRUE,0,0,FALSE,7000,1,0,0,0,1,4,1952,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,200,4,2065,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
+2,3,6,FALSE,0,5500,5000,TRUE,25,FALSE,TRUE,0,0,FALSE,2500,1,0,0,0,1,4,1985,1,1,0,0,0,7,6.2776,FALSE,FALSE,0,0,0,201,5,2066,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2018
+2,4,7,FALSE,0,5500,300,TRUE,20,FALSE,FALSE,0,0,FALSE,2300,1,0,0,0,0,4,1982,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,202,6,2067,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
+2,4,8,FALSE,0,5500,600,TRUE,20,FALSE,FALSE,0,0,FALSE,1000,1,0,0,0,0,4,1971,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,203,7,2068,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
+3,5,9,TRUE,0,5500,900,TRUE,20,FALSE,FALSE,0,0,FALSE,2000,1,0,0,0,0,1,1968,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,204,8,2069,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
+4,6,10,TRUE,0,5500,1200,FALSE,20,FALSE,TRUE,0,0,FALSE,3000,12,5,0,0,1,1,1964,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,205,9,2070,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
+5,7,11,TRUE,0,5500,1500,FALSE,20,FALSE,TRUE,0,0,FALSE,4000,1,0,0,20,1,1,1967,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,206,10,2071,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
+6,8,12,TRUE,0,5500,8000,FALSE,20,FALSE,TRUE,0,0,FALSE,7000,1,0,0,0,1,1,1951,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,207,11,2072,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2010
+7,9,13,TRUE,0,5500,700,FALSE,20,FALSE,FALSE,0,0,FALSE,2500,1,0,0,0,1,1,1960,1,1,0,0,0,2,1.7936,FALSE,FALSE,0,0,0,208,12,2073,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2002
+8,10,14,TRUE,0,5500,0,FALSE,40,TRUE,TRUE,2500,0,FALSE,2300,1,0,0,0,0,1,1959,1,1,0,0,0,22,19.7296,FALSE,FALSE,0,0,0,0,13,2074,FALSE,FALSE,0,FALSE,400,50,FALSE,80,3,3,50,2019
