+++ gettsim/tests/test_kinderzuschlag.py
@@ -32,7 +32,7 @@ INPUT_COLS = [
     "unterhaltsvors_m",
     "jahr",
 ]
-OUT_COLS = ["kinderzuschlag_temp"]
+OUT_COLS = ["_kinderzuschlag_m_vorläufig"]
 YEARS = [2006, 2009, 2011, 2013, 2016, 2017, 2019, 2020]
